# contracts/openapi-v1.json
{
  "components": {
    "schemas": {
      "CaseResponse": {
        "additionalProperties": false,
        "properties": {
          "description": {
            "title": "Description",
            "type": "string"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "path_id": {
            "$ref": "#/components/schemas/PathId"
          },
          "status": {
            "$ref": "#/components/schemas/CaseStatus"
          },
          "version": {
            "title": "Version",
            "type": "string"
          }
        },
        "required": [
          "id",
          "version",
          "path_id",
          "name",
          "description",
          "status"
        ],
        "title": "CaseResponse",
        "type": "object"
      },
      "CaseStatus": {
        "enum": [
          "OPEN",
          "LOCKED",
          "CLOSED"
        ],
        "title": "CaseStatus",
        "type": "string"
      },
      "CatalogueSectionResponse": {
        "additionalProperties": false,
        "properties": {
          "cases": {
            "items": {
              "$ref": "#/components/schemas/CaseResponse"
            },
            "title": "Cases",
            "type": "array"
          },
          "path": {
            "$ref": "#/components/schemas/PathResponse"
          }
        },
        "required": [
          "path",
          "cases"
        ],
        "title": "CatalogueSectionResponse",
        "type": "object"
      },
      "CatalogueSectionsResponse": {
        "additionalProperties": false,
        "properties": {
          "sections": {
            "items": {
              "$ref": "#/components/schemas/PathResponse"
            },
            "title": "Sections",
            "type": "array"
          }
        },
        "required": [
          "sections"
        ],
        "title": "CatalogueSectionsResponse",
        "type": "object"
      },
      "ComicKind": {
        "enum": [
          "OPENING",
          "CLOSING"
        ],
        "title": "ComicKind",
        "type": "string"
      },
      "ComicPageResponse": {
        "additionalProperties": false,
        "properties": {
          "alt_text": {
            "title": "Alt Text",
            "type": "string"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "image_url": {
            "title": "Image Url",
            "type": "string"
          },
          "narration": {
            "title": "Narration",
            "type": "string"
          },
          "position": {
            "title": "Position",
            "type": "integer"
          },
          "title": {
            "title": "Title",
            "type": "string"
          }
        },
        "required": [
          "id",
          "position",
          "title",
          "narration",
          "image_url",
          "alt_text"
        ],
        "title": "ComicPageResponse",
        "type": "object"
      },
      "ComicSequenceResponse": {
        "additionalProperties": false,
        "properties": {
          "case_id": {
            "title": "Case Id",
            "type": "string"
          },
          "case_version": {
            "title": "Case Version",
            "type": "string"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "kind": {
            "$ref": "#/components/schemas/ComicKind"
          },
          "pages": {
            "items": {
              "$ref": "#/components/schemas/ComicPageResponse"
            },
            "title": "Pages",
            "type": "array"
          },
          "skippable": {
            "title": "Skippable",
            "type": "boolean"
          }
        },
        "required": [
          "id",
          "case_id",
          "case_version",
          "kind",
          "skippable",
          "pages"
        ],
        "title": "ComicSequenceResponse",
        "type": "object"
      },
      "CreateRoundRequest": {
        "additionalProperties": false,
        "properties": {
          "case_id": {
            "maxLength": 64,
            "minLength": 1,
            "title": "Case Id",
            "type": "string"
          },
          "path_id": {
            "maxLength": 64,
            "minLength": 1,
            "title": "Path Id",
            "type": "string"
          },
          "player_id": {
            "default": "demo-hercule",
            "maxLength": 100,
            "minLength": 1,
            "title": "Player Id",
            "type": "string"
          }
        },
        "required": [
          "path_id",
          "case_id"
        ],
        "title": "CreateRoundRequest",
        "type": "object"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
      "LiveHealthResponse": {
        "additionalProperties": false,
        "properties": {
          "status": {
            "default": "alive",
            "title": "Status",
            "type": "string"
          }
        },
        "title": "LiveHealthResponse",
        "type": "object"
      },
      "OpeningResponse": {
        "additionalProperties": false,
        "properties": {
          "case": {
            "$ref": "#/components/schemas/CaseResponse"
          },
          "round": {
            "$ref": "#/components/schemas/RoundResponse"
          },
          "sequence": {
            "$ref": "#/components/schemas/ComicSequenceResponse"
          }
        },
        "required": [
          "round",
          "case",
          "sequence"
        ],
        "title": "OpeningResponse",
        "type": "object"
      },
      "PathId": {
        "enum": [
          "DETECTIVE_ACADEMY",
          "PUPPY",
          "ADULT_DOG",
          "SENIOR_DOG"
        ],
        "title": "PathId",
        "type": "string"
      },
      "PathResponse": {
        "additionalProperties": false,
        "properties": {
          "access_message": {
            "title": "Access Message",
            "type": "string"
          },
          "description": {
            "title": "Description",
            "type": "string"
          },
          "id": {
            "$ref": "#/components/schemas/PathId"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "ranked": {
            "title": "Ranked",
            "type": "boolean"
          },
          "status": {
            "$ref": "#/components/schemas/PathStatus"
          }
        },
        "required": [
          "id",
          "name",
          "description",
          "ranked",
          "status",
          "access_message"
        ],
        "title": "PathResponse",
        "type": "object"
      },
      "PathStatus": {
        "enum": [
          "OPEN",
          "COMING_SOON",
          "LOCKED"
        ],
        "title": "PathStatus",
        "type": "string"
      },
      "ReadyHealthResponse": {
        "additionalProperties": false,
        "properties": {
          "checks": {
            "additionalProperties": {
              "type": "string"
            },
            "title": "Checks",
            "type": "object"
          },
          "status": {
            "title": "Status",
            "type": "string"
          }
        },
        "required": [
          "status",
          "checks"
        ],
        "title": "ReadyHealthResponse",
        "type": "object"
      },
      "RoundResponse": {
        "additionalProperties": false,
        "properties": {
          "case_id": {
            "title": "Case Id",
            "type": "string"
          },
          "case_version": {
            "title": "Case Version",
            "type": "string"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "intro_completed_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Intro Completed At"
          },
          "path_id": {
            "$ref": "#/components/schemas/PathId"
          },
          "player_id": {
            "title": "Player Id",
            "type": "string"
          },
          "started_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Started At"
          },
          "status": {
            "$ref": "#/components/schemas/RoundStatus"
          }
        },
        "required": [
          "id",
          "player_id",
          "path_id",
          "case_id",
          "case_version",
          "status",
          "created_at",
          "started_at",
          "intro_completed_at"
        ],
        "title": "RoundResponse",
        "type": "object"
      },
      "RoundStatus": {
        "enum": [
          "CREATED",
          "INTRO_PENDING",
          "ACTIVE"
        ],
        "title": "RoundStatus",
        "type": "string"
      },
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
      "VersionResponse": {
        "additionalProperties": false,
        "properties": {
          "build_version": {
            "title": "Build Version",
            "type": "string"
          },
          "contract_version": {
            "title": "Contract Version",
            "type": "string"
          },
          "environment": {
            "title": "Environment",
            "type": "string"
          },
          "runtime_role": {
            "title": "Runtime Role",
            "type": "string"
          }
        },
        "required": [
          "build_version",
          "contract_version",
          "runtime_role",
          "environment"
        ],
        "title": "VersionResponse",
        "type": "object"
      },
      "WorkspaceResponse": {
        "additionalProperties": false,
        "properties": {
          "board_message": {
            "title": "Board Message",
            "type": "string"
          },
          "case": {
            "$ref": "#/components/schemas/CaseResponse"
          },
          "evidence_count": {
            "title": "Evidence Count",
            "type": "integer"
          },
          "round": {
            "$ref": "#/components/schemas/RoundResponse"
          },
          "suspect_count": {
            "title": "Suspect Count",
            "type": "integer"
          }
        },
        "required": [
          "round",
          "case",
          "board_message",
          "evidence_count",
          "suspect_count"
        ],
        "title": "WorkspaceResponse",
        "type": "object"
      }
    }
  },
  "info": {
    "title": "Fraud Graph Arena",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
    "/api/v1/catalogue/sections": {
      "get": {
        "operationId": "list_sections_api_v1_catalogue_sections_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CatalogueSectionsResponse"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "summary": "List Sections",
        "tags": [
          "catalogue"
        ]
      }
    },
    "/api/v1/catalogue/{section}": {
      "get": {
        "operationId": "get_section_api_v1_catalogue__section__get",
        "parameters": [
          {
            "in": "path",
            "name": "section",
            "required": true,
            "schema": {
              "title": "Section",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CatalogueSectionResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Section",
        "tags": [
          "catalogue"
        ]
      }
    },
    "/api/v1/health/live": {
      "get": {
        "operationId": "live_api_v1_health_live_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LiveHealthResponse"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "summary": "Live",
        "tags": [
          "health"
        ]
      }
    },
    "/api/v1/health/ready": {
      "get": {
        "operationId": "ready_api_v1_health_ready_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReadyHealthResponse"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "summary": "Ready",
        "tags": [
          "health"
        ]
      }
    },
    "/api/v1/health/version": {
      "get": {
        "operationId": "version_api_v1_health_version_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/VersionResponse"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "summary": "Version",
        "tags": [
          "health"
        ]
      }
    },
    "/api/v1/rounds": {
      "post": {
        "operationId": "create_round_api_v1_rounds_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateRoundRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RoundResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Round",
        "tags": [
          "rounds"
        ]
      }
    },
    "/api/v1/rounds/{round_id}/opening": {
      "get": {
        "operationId": "opening_api_v1_rounds__round_id__opening_get",
        "parameters": [
          {
            "in": "path",
            "name": "round_id",
            "required": true,
            "schema": {
              "title": "Round Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OpeningResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Opening",
        "tags": [
          "rounds"
        ]
      }
    },
    "/api/v1/rounds/{round_id}/opening/complete": {
      "post": {
        "operationId": "complete_opening_api_v1_rounds__round_id__opening_complete_post",
        "parameters": [
          {
            "in": "path",
            "name": "round_id",
            "required": true,
            "schema": {
              "title": "Round Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RoundResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Complete Opening",
        "tags": [
          "rounds"
        ]
      }
    },
    "/api/v1/rounds/{round_id}/start": {
      "post": {
        "operationId": "start_round_api_v1_rounds__round_id__start_post",
        "parameters": [
          {
            "in": "path",
            "name": "round_id",
            "required": true,
            "schema": {
              "title": "Round Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RoundResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Start Round",
        "tags": [
          "rounds"
        ]
      }
    },
    "/api/v1/rounds/{round_id}/workspace": {
      "get": {
        "operationId": "workspace_api_v1_rounds__round_id__workspace_get",
        "parameters": [
          {
            "in": "path",
            "name": "round_id",
            "required": true,
            "schema": {
              "title": "Round Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WorkspaceResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Workspace",
        "tags": [
          "rounds"
        ]
      }
    }
  }
}